requirements.txt
flask
peewee
docker
pytest
swagger-ui-py
